+++ joulescope_ui/logging_util.py
@@ -19,8 +19,12 @@ import logging
 from logging import FileHandler
 import time
 import datetime
+import json
 import os
+import sys
 from joulescope_ui.config import APP_PATH
+from . import VERSION as UI_VERSION
+from joulescope import VERSION as DRIVER_VERSION
 
 
 LOG_PATH = os.path.join(APP_PATH, 'log')
@@ -42,6 +46,43 @@ for name, value in list(LEVELS.items()):
     LEVELS[value] = value
 assert(logging.CRITICAL == 50)
 assert(logging.DEBUG == 10)
+
+
+_BANNER = """\
+Joulescope User Interface
+UI Version = {ui_version}
+Driver Version = {driver_version}"""
+
+
+def _make_banner():
+    banner = _BANNER.format(ui_version=UI_VERSION, driver_version=DRIVER_VERSION)
+    lines = banner.split('\n')
+    line_length = max([len(x) for x in lines]) + 4
+    lines = ['* ' + line + (' ' * (line_length - len(line) - 3)) + '*' for line in lines]
+    k = '*' * line_length
+    lines = [k] + lines + [k, '']
+    return '\n'.join(lines)
+
+
+def _make_info():
+    frozen = getattr(sys, 'frozen', False)
+    if frozen:
+        frozen = getattr(sys, '_MEIPASS', frozen)
+    info = {
+        'joulescope': {
+            'ui_version': UI_VERSION,
+            'driver_version': DRIVER_VERSION,
+        },
+        'platform': {
+            'name': sys.platform,
+            'python_version': sys.version,
+            'executable': sys.executable,
+            'frozen': frozen,
+            'app_path': APP_PATH,
+            'log_path': LOG_PATH,
+        }
+    }
+    return json.dumps(info, indent=2)
 
 
 def _cleanup_logfiles():
@@ -66,6 +107,8 @@ def logging_config(stream_log_level=None, file_log_level=None):
     :param file_log_level: The logging level for the log file which
         can be the integer value or name.  None (default) is 'INFO'.
     """
+    banner = _make_banner()
+    banner = banner + '\ninfo = ' + _make_info() + '\n'
     os.makedirs(LOG_PATH, exist_ok=True)
     d = datetime.datetime.utcnow()
     time_str = d.strftime('%Y%m%d_%H%M%S')
@@ -74,6 +117,7 @@ def logging_config(stream_log_level=None, file_log_level=None):
     stream_lvl = logging.WARNING if stream_log_level is None else LEVELS[stream_log_level]
     stream_fmt = logging.Formatter(STREAM_VERBOSE_FMT)
     stream_hnd = logging.StreamHandler()
+    stream_hnd.stream.write(banner)
     stream_hnd.setFormatter(stream_fmt)
     stream_hnd.setLevel(stream_lvl)
 
@@ -81,6 +125,7 @@ def logging_config(stream_log_level=None, file_log_level=None):
     if file_lvl < LEVELS['OFF']:
         file_fmt = logging.Formatter(FILE_FMT)
         file_hnd = FileHandler(filename=filename)
+        file_hnd.stream.write(banner)
         file_hnd.setFormatter(file_fmt)
         file_hnd.setLevel(file_lvl)
 
@@ -91,3 +136,4 @@ def logging_config(stream_log_level=None, file_log_level=None):
 
     root_log.setLevel(min([stream_lvl, file_lvl]))
     _cleanup_logfiles()
+    root_log.info('logging configuration: stream_level=%s, file_level=%s', stream_lvl, file_lvl)